Adds roundstart and roundend animations

+++ code/controllers/subsystem/ticker.dm
@@ -379,22 +379,41 @@ var/datum/subsystem/ticker/ticker
 
 
 /datum/subsystem/ticker/proc/create_characters()
+	var/list/livings = list()
 	for(var/mob/new_player/player in player_list)
 		if(player.ready && player.mind)
 			joined_player_list += player.ckey
 			if(player.mind.assigned_role=="AI")
 				player.close_spawn_windows()
-				player.AIize()
+				var/mob/living = player.AIize()
+				if(living)
+					living.notransform = TRUE
+					if(living.client)
+						var/obj/screen/splash/S = new(living.client, TRUE)
+						S.Fade(TRUE)
+					livings += living
 			else
-				player.create_character()
-				qdel(player)
+				var/mob/living = player.create_character()
+				if(living)
+					qdel(player)
+					living.notransform = TRUE
+					if(living.client)
+						var/obj/screen/splash/S = new(living.client, TRUE)
+						S.Fade(TRUE)
+					livings += living
 		else
 			if(player.client)
 				if(player.client.prefs.agree < MAXAGREE)
 					player.disclaimer()
 				else
 					player.new_player_panel()
+	if(livings.len)
+		addtimer(src, "release_characters", 30, FALSE, livings)
 
+/datum/subsystem/ticker/proc/release_characters(list/livings)
+	for(var/I in livings)
+		var/mob/living/L = I
+		L.notransform = FALSE
 
 /datum/subsystem/ticker/proc/collect_minds()
 	for(var/mob/living/player in player_list)
